agents/Xpander_Travel_Assistant/requirements.txt
xpander-sdk
agno
python-dotenv
asyncio
python-dateutil
slack-sdk
certifi
requests
openai
sqlalchemy
xpander-utils
